+++ Changes.md
@@ -74,6 +74,7 @@ API
     - PresetsPlugValueWidget
 - SetAlgo : Added Python binding for `affectsSetExpression()`.
 - Shader : Added `affectsAttributes()` protected method.
+- MessageWidget : Added MessageSummaryWidget class to simplify the display of message counts in other UIs.
 
 Breaking Changes
 ----------------